pyproject.toml
[project]
name = "data-lakehouse-olist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.93",
    "duckdb>=1.1.3",
    "kagglehub>=0.3.6",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=18.1.0",
    "python-dotenv>=1.0.1",
]
